+++ packages/iov-bns/src/decode.spec.ts
@@ -44,7 +44,6 @@ import {
   isRegisterUsernameTx,
   isReleaseEscrowTx,
   isReturnEscrowTx,
-  isTallyTx,
   isUpdateEscrowPartiesTx,
   isUpdateMultisignatureTx,
   isUpdateTargetsOfUsernameTx,
@@ -67,7 +66,7 @@ describe("Decode", () => {
       targets: [
         {
           blockchainId: "wonderland",
-          address: toUtf8("12345W"),
+          address: "12345W",
         },
       ],
     };
@@ -400,7 +399,7 @@ describe("Decode", () => {
           newTargets: [
             {
               blockchainId: "wonderland",
-              address: toUtf8("0xAABB001122DD"),
+              address: "0xAABB001122DD",
             },
           ],
         },
@@ -425,11 +424,11 @@ describe("Decode", () => {
           targets: [
             {
               blockchainId: "chain1",
-              address: toUtf8("23456782367823X"),
+              address: "23456782367823X",
             },
             {
               blockchainId: "chain2",
-              address: toUtf8("0x001100aabbccddffeeddaa8899776655"),
+              address: "0x001100aabbccddffeeddaa8899776655",
             },
           ],
         },
@@ -805,20 +804,6 @@ describe("Decode", () => {
         throw new Error("unexpected transaction kind");
       }
       expect(parsed.selection).toEqual(VoteOption.Yes);
-      expect(parsed.proposalId).toEqual(187723859034111);
-    });
-
-    it("works for TallyTx", () => {
-      const transactionMessage: codecImpl.bnsd.ITx = {
-        govTallyMsg: {
-          metadata: { schema: 1 },
-          proposalId: fromHex("aabbddeeffff"),
-        },
-      };
-      const parsed = parseMsg(defaultBaseTx, transactionMessage);
-      if (!isTallyTx(parsed)) {
-        throw new Error("unexpected transaction kind");
-      }
       expect(parsed.proposalId).toEqual(187723859034111);
     });
   });