+++ packages/iov-bns/src/encode.ts
@@ -33,7 +33,6 @@ import {
   RegisterUsernameTx,
   ReleaseEscrowTx,
   ReturnEscrowTx,
-  TallyTx,
   UpdateEscrowPartiesTx,
   UpdateMultisignatureTx,
   UpdateTargetsOfUsernameTx,
@@ -42,8 +41,6 @@ import {
   VoteTx,
 } from "./types";
 import { decodeBnsAddress, identityToAddress } from "./util";
-
-const { toUtf8 } = Encoding;
 
 function encodeInt(intNumber: number): number | null {
   if (!Number.isInteger(intNumber)) {
@@ -175,7 +172,7 @@ function buildSwapClaimTx(tx: SwapClaimTransaction): codecImpl.bnsd.ITx {
 
 function buildSwapAbortTransaction(tx: SwapAbortTransaction): codecImpl.bnsd.ITx {
   return {
-    aswapReturnMsg: codecImpl.aswap.ReturnSwapMsg.create({
+    aswapReturnMsg: codecImpl.aswap.ReturnMsg.create({
       metadata: { schema: 1 },
       swapId: tx.swapId.data,
     }),
@@ -187,7 +184,7 @@ function buildSwapAbortTransaction(tx: SwapAbortTransaction): codecImpl.bnsd.ITx
 function encodeChainAddressPair(pair: ChainAddressPair): codecImpl.username.IBlockchainAddress {
   return {
     blockchainId: pair.chainId,
-    address: toUtf8(pair.address),
+    address: pair.address,
   };
 }
 
@@ -384,15 +381,6 @@ function buildVoteTx(tx: VoteTx): codecImpl.bnsd.ITx {
   };
 }
 
-function buildTallyTx(tx: TallyTx): codecImpl.bnsd.ITx {
-  return {
-    govTallyMsg: {
-      metadata: { schema: 1 },
-      proposalId: encodeNumericId(tx.proposalId),
-    },
-  };
-}
-
 export function buildMsg(tx: UnsignedTransaction): codecImpl.bnsd.ITx {
   if (!isBnsTx(tx)) {
     throw new Error("Transaction is not a BNS transaction");
@@ -438,8 +426,6 @@ export function buildMsg(tx: UnsignedTransaction): codecImpl.bnsd.ITx {
       return buildCreateProposalTx(tx);
     case "bns/vote":
       return buildVoteTx(tx);
-    case "bns/tally":
-      return buildTallyTx(tx);
 
     default:
       throw new Error("Received transaction of unsupported kind.");