+++ CHANGELOG.md
@@ -14,12 +14,13 @@ Breaking changes
 
 - @iov/bns: `ElectionRule`, `Proposal` and `VersionedId` all use numeric IDs now
   to conform with `Electorate`.
-- @iov/bns: Remove support for weave 0.16.x; add support for weave 0.18.x.
-- @iov/bns: Remove `AddAddressToUsernameTx` and `RemoveAddressFromUsernameTx` in
+- @iov/bns: Remove support for weave 0.16.x; add support for weave 0.19.x.
+- @iov/bns: Remove `AddAddressToUsernameTx` and `UpdateTargetsOfUsernameTx` in
   favour of the new `UpdateTargetsOfUsernameTx`.
 - @iov/bns: Rename `BnsUsernameNft.addresses` to `.targets`.
 - @iov/bns: Rename `RegisterUsernameTx.addresses` to `.targets`.
 - @iov/bns: `RegisterUsernameTx.username` now has to include a \*iov suffix.
+- @iov/bns: Remove `TallyTx` as tallies will now be performed automatically.
 - @iov/core: Rename package to @iov/multichain.
 
 ## 0.15.1